Deepgram/deepgram_handler/pyproject.toml
[project]
name = "twilio-proxy"
version = "0.1.0"
description = "Twilio proxy - connects calls to Deepgram Voice Agent, replies Hi to SMS"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "websockets>=13.0",
  "python-multipart>=0.0.12",
  "httpx>=0.28.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
